Fix typo: Infered -> inferred (#17417)
