--- tests/baselines/reference/declarationEmitInferedTypeAlias8.symbols
@@ -1,19 +1,0 @@
-=== tests/cases/compiler/declarationEmitInferedTypeAlias8.ts ===
-type Foo<T> = T | { x: Foo<T> };
->Foo : Symbol(Foo, Decl(declarationEmitInferedTypeAlias8.ts, 0, 0))
->T : Symbol(T, Decl(declarationEmitInferedTypeAlias8.ts, 0, 9))
->T : Symbol(T, Decl(declarationEmitInferedTypeAlias8.ts, 0, 9))
->x : Symbol(x, Decl(declarationEmitInferedTypeAlias8.ts, 0, 19))
->Foo : Symbol(Foo, Decl(declarationEmitInferedTypeAlias8.ts, 0, 0))
->T : Symbol(T, Decl(declarationEmitInferedTypeAlias8.ts, 0, 9))
-
-var x: Foo<number[]>;
->x : Symbol(x, Decl(declarationEmitInferedTypeAlias8.ts, 1, 3))
->Foo : Symbol(Foo, Decl(declarationEmitInferedTypeAlias8.ts, 0, 0))
-
-function returnSomeGlobalValue() {
->returnSomeGlobalValue : Symbol(returnSomeGlobalValue, Decl(declarationEmitInferedTypeAlias8.ts, 1, 21))
-
-    return x;
->x : Symbol(x, Decl(declarationEmitInferedTypeAlias8.ts, 1, 3))
-}
