+++ tests/baselines/reference/declarationEmitInferredDefaultExportType.types
@@ -1,4 +1,4 @@
-=== tests/cases/compiler/declarationEmitInferedDefaultExportType.ts ===
+=== tests/cases/compiler/declarationEmitInferredDefaultExportType.ts ===
 // test.ts
 export default {
 >{  foo: [],  bar: undefined,  baz: null} : { foo: undefined[]; bar: undefined; baz: null; }
