+++ tests/baselines/reference/declarationEmitInferredDefaultExportType2.symbols
@@ -1,0 +1,13 @@
+=== tests/cases/compiler/declarationEmitInferredDefaultExportType2.ts ===
+// test.ts
+export = {
+  foo: [],
+>foo : Symbol(foo, Decl(declarationEmitInferredDefaultExportType2.ts, 1, 10))
+
+  bar: undefined,
+>bar : Symbol(bar, Decl(declarationEmitInferredDefaultExportType2.ts, 2, 10))
+>undefined : Symbol(undefined)
+
+  baz: null
+>baz : Symbol(baz, Decl(declarationEmitInferredDefaultExportType2.ts, 3, 17))
+}
