+++ tests/baselines/reference/declarationEmitInferredDefaultExportType2.types
@@ -1,4 +1,4 @@
-=== tests/cases/compiler/declarationEmitInferedDefaultExportType2.ts ===
+=== tests/cases/compiler/declarationEmitInferredDefaultExportType2.ts ===
 // test.ts
 export = {
 >{  foo: [],  bar: undefined,  baz: null} : { foo: undefined[]; bar: undefined; baz: null; }
