+++ tests/baselines/reference/declarationEmitInferredTypeAlias1.symbols
