+++ tests/baselines/reference/declarationEmitInferredTypeAlias1.types
