+++ tests/baselines/reference/declarationEmitInferredTypeAlias2.symbols
