+++ tests/baselines/reference/declarationEmitInferredTypeAlias2.types
