+++ tests/baselines/reference/declarationEmitInferredTypeAlias3.symbols
