+++ tests/baselines/reference/declarationEmitInferredTypeAlias3.types
