+++ tests/baselines/reference/declarationEmitInferredTypeAlias4.symbols
@@ -1,0 +1,21 @@
+=== tests/cases/compiler/declarationEmitInferredTypeAlias4.ts ===
+function f<A>() {
+>f : Symbol(f, Decl(declarationEmitInferredTypeAlias4.ts, 0, 0))
+>A : Symbol(A, Decl(declarationEmitInferredTypeAlias4.ts, 0, 11))
+
+    type Foo<T> = T | { x: Foo<T> };
+>Foo : Symbol(Foo, Decl(declarationEmitInferredTypeAlias4.ts, 0, 17))
+>T : Symbol(T, Decl(declarationEmitInferredTypeAlias4.ts, 1, 13))
+>T : Symbol(T, Decl(declarationEmitInferredTypeAlias4.ts, 1, 13))
+>x : Symbol(x, Decl(declarationEmitInferredTypeAlias4.ts, 1, 23))
+>Foo : Symbol(Foo, Decl(declarationEmitInferredTypeAlias4.ts, 0, 17))
+>T : Symbol(T, Decl(declarationEmitInferredTypeAlias4.ts, 1, 13))
+
+    var x: Foo<A[]>;
+>x : Symbol(x, Decl(declarationEmitInferredTypeAlias4.ts, 2, 7))
+>Foo : Symbol(Foo, Decl(declarationEmitInferredTypeAlias4.ts, 0, 17))
+>A : Symbol(A, Decl(declarationEmitInferredTypeAlias4.ts, 0, 11))
+
+    return x;
+>x : Symbol(x, Decl(declarationEmitInferredTypeAlias4.ts, 2, 7))
+}
