+++ tests/baselines/reference/declarationEmitInferredTypeAlias4.types
@@ -1,4 +1,4 @@
-=== tests/cases/compiler/declarationEmitInferedTypeAlias4.ts ===
+=== tests/cases/compiler/declarationEmitInferredTypeAlias4.ts ===
 function f<A>() {
 >f : <A>() => A[] | { x: A[] | any; }
 >A : A
