+++ tests/baselines/reference/declarationEmitInferredTypeAlias5.js
@@ -1,10 +1,12 @@
-//// [tests/cases/compiler/declarationEmitInferedTypeAlias7.ts] ////
+//// [tests/cases/compiler/declarationEmitInferredTypeAlias5.ts] ////
 
 //// [0.ts]
 export type Data = string | boolean;
 let obj: Data = true;
 
 //// [1.ts]
+import * as Z from "./0"
+//let v2: Z.Data;
 let v = "str" || true;
 export { v }
 
@@ -15,6 +17,7 @@ var obj = true;
 //// [1.js]
 "use strict";
 exports.__esModule = true;
+//let v2: Z.Data;
 var v = "str" || true;
 exports.v = v;
 
@@ -22,5 +25,6 @@ exports.v = v;
 //// [0.d.ts]
 export declare type Data = string | boolean;
 //// [1.d.ts]
-declare let v: string | boolean;
+import * as Z from "./0";
+declare let v: Z.Data;
 export { v };
