+++ tests/baselines/reference/declarationEmitInferredTypeAlias5.symbols
