+++ tests/baselines/reference/declarationEmitInferredTypeAlias5.types
