+++ tests/baselines/reference/declarationEmitInferredTypeAlias6.symbols
