+++ tests/baselines/reference/declarationEmitInferredTypeAlias6.types
