+++ tests/baselines/reference/declarationEmitInferredTypeAlias7.symbols
