+++ tests/baselines/reference/declarationEmitInferredTypeAlias7.types
