+++ tests/baselines/reference/declarationEmitInferredTypeAlias8.js
@@ -1,4 +1,4 @@
-//// [declarationEmitInferedTypeAlias8.ts]
+//// [declarationEmitInferredTypeAlias8.ts]
 type Foo<T> = T | { x: Foo<T> };
 var x: Foo<number[]>;
 
@@ -6,14 +6,14 @@ function returnSomeGlobalValue() {
     return x;
 }
 
-//// [declarationEmitInferedTypeAlias8.js]
+//// [declarationEmitInferredTypeAlias8.js]
 var x;
 function returnSomeGlobalValue() {
     return x;
 }
 
 
-//// [declarationEmitInferedTypeAlias8.d.ts]
+//// [declarationEmitInferredTypeAlias8.d.ts]
 declare type Foo<T> = T | {
     x: Foo<T>;
 };
