+++ tests/baselines/reference/declarationEmitInferredTypeAlias8.types
@@ -1,4 +1,4 @@
-=== tests/cases/compiler/declarationEmitInferedTypeAlias8.ts ===
+=== tests/cases/compiler/declarationEmitInferredTypeAlias8.ts ===
 type Foo<T> = T | { x: Foo<T> };
 >Foo : Foo<T>
 >T : T
