+++ tests/cases/compiler/declarationEmitInferredDefaultExportType.ts
