+++ tests/cases/compiler/declarationEmitInferredDefaultExportType2.ts
