+++ tests/cases/compiler/declarationEmitInferredTypeAlias1.ts
