+++ tests/cases/compiler/declarationEmitInferredTypeAlias2.ts
