+++ tests/cases/compiler/declarationEmitInferredTypeAlias3.ts
