+++ tests/cases/compiler/declarationEmitInferredTypeAlias4.ts
