+++ tests/cases/compiler/declarationEmitInferredTypeAlias5.ts
