+++ tests/cases/compiler/declarationEmitInferredTypeAlias6.ts
