+++ tests/cases/compiler/declarationEmitInferredTypeAlias7.ts
