+++ tests/cases/compiler/declarationEmitInferredTypeAlias8.ts
