+++ tests/cases/compiler/declarationEmitInferredTypeAlias9.ts
